% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_maf_from_geno_data.R
\name{calculate_maf_from_geno_data}
\alias{calculate_maf_from_geno_data}
\title{Calculate the minor allele frequency from control genotype data in
num_controls by num_snp format.}
\usage{
calculate_maf_from_geno_data(control_geno_data)
}
\arguments{
\item{control_geno_data}{Control genotype data.}
}
\value{
A vector of minor allele frequencies of length equal to the number of SNPs.
}
\description{
Calculate the minor allele frequency from control genotype data in
num_controls by num_snp format.
}
