% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_maf_from_haplo_data.R
\name{calculate_maf_from_haplo_data}
\alias{calculate_maf_from_haplo_data}
\title{Calculate the MAF from a single Hapgen2 control haplotype data file.}
\usage{
calculate_maf_from_haplo_data(control_haplo_file)
}
\arguments{
\item{control_haplo_file}{A file containing Hapgen2 control haplotype data.}
}
\value{
A vector of minor allele frequencies of length equal to the number of SNPs.
}
\description{
Calculate the MAF from a single Hapgen2 control haplotype data file.
}
